refactor(submodel-view-object-util):move functions if only used in one component (#68)
